refactor: rename open-agents-agent to agent
