+++ packages/agent/types.ts
@@ -1,7 +1,7 @@
 import type { SandboxState } from "@open-agents/sandbox";
 import type { LanguageModel } from "ai";
 import { z } from "zod";
-import type { AgentSandboxContext } from "./open-agents-agent";
+import type { AgentSandboxContext } from "./agent";
 import type { SkillMetadata } from "./skills/types";
 
 export const todoStatusSchema = z.enum(["pending", "in_progress", "completed"]);
